moved vector unit tests outside of tests folder
