+++ eckity/genetic_encodings/ga/bit_string_vector.py
@@ -15,24 +15,31 @@ class BitStringVector(Vector):
     Parameters
     ----------
     fitness : Fitness
-        Fitness handler class, responsible of keeping the fitness value of the individual.
+        Fitness handler class.
+        Responsible of keeping the fitness value of the individual.
 
     length : int
         Vector length - the number of cells in the vector.
 
     bounds : tuple or list of tuples
-        Min/Max values for each vector cell (if of length n), or the minimum and maximum (if of length 1).
+        Min/Max values for each vector cell (if of length n),
+        or the minimum and maximum (if of length 1).
     """
 
     def __init__(self,
                  fitness,
                  length,
-                 bounds=(0, 1)):
-        super().__init__(fitness=fitness, length=length, bounds=bounds)
+                 bounds=(0, 1),
+                 vector=None):
+        super().__init__(fitness=fitness,
+                         length=length,
+                         bounds=bounds,
+                         vector=vector)
 
     def get_random_number_in_bounds(self, index):
         """
-        Return a random number of available cell values - 0 or 1, with equal probability.
+        Return a random number of available cell values (0 or 1),
+        with equal probability.
 
         Parameters
         ----------
@@ -44,23 +51,13 @@ class BitStringVector(Vector):
         int
             random value according to bounds field
         """
-        # todo check if need to check bounds - is it tuple always?
         return randint(self.bounds[0], self.bounds[1])
 
     def bit_flip(self, index):
         """
-        Return a random number of available cell values - 0 or 1, with equal probability.
-
-        Parameters
-        ----------
-        index : int
-            cell index
-
-        Returns
-        -------
-        int
-            random value according to bounds field
+        Flip the bit in the given index.
         """
-        return self.bounds[1] if self.cell_value(index) == self.bounds[0] else self.bounds[0]
+        return self.bounds[1] \
+            if self.cell_value(index) == self.bounds[0] else self.bounds[0]
 
 # end class bit string vector
